LPS-70397 Move into `auth.publicpath` package
